requirements.txt
numpy==2.2.1
pandas==2.2.3
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.2
timm==1.0.13
torch==2.2.1
torchvision==0.17.1
tqdm==4.67.1
